// src/interfaces/IItemComboBaseDTO.ts
export interface IItemComboBaseDTO {
    key: string;
    text: string;
    itemType?: number | 0;
    disabled?: boolean | false;
    codigo?: string;
    descripcion?: string;
    manoObraTipoId?: string;
    nombre?: string;
    catalogoAccionesCodigo?: string;
    areaReporteNombre?: string;
    contratoId?: string;
  }
  